+++ react/features/toolbar/components/index.js
@@ -1,2 +1,1 @@
 export { default as Toolbar } from './Toolbar';
-export { default as ExtendedToolbar } from './ExtendedToolbar';
